separate controllers into hrl controllers file
